Simplify source folder structure in MaterialXRender
