+++ source/MaterialXRender/ExceptionShaderValidationError.h
