+++ source/MaterialXRender/ImageHandler.h
