+++ source/MaterialXRender/LightHandler.h
