+++ source/MaterialXRender/Mesh.h
