+++ source/MaterialXRender/OiioImageLoader.cpp
