+++ source/MaterialXRender/SampleObjLoader.h
@@ -11,7 +11,7 @@
 
 #include <string>
 #include <memory>
-#include <MaterialXRender/Handlers/GeometryHandler.h>
+#include <MaterialXRender/GeometryHandler.h>
 
 namespace MaterialX
 {
