+++ source/MaterialXRender/ViewHandler.h
